+++ apps/indexer/lib/indexer/application.ex
@@ -24,7 +24,13 @@ defmodule Indexer.Application do
 
     json_rpc_named_arguments = Application.fetch_env!(:indexer, :json_rpc_named_arguments)
 
+    pool_size =
+      Application.get_env(:indexer, Indexer.Fetcher.TokenInstance.Retry)[:concurrency] +
+        Application.get_env(:indexer, Indexer.Fetcher.TokenInstance.Realtime)[:concurrency] +
+        Application.get_env(:indexer, Indexer.Fetcher.TokenInstance.Sanitize)[:concurrency]
+
     base_children = [
+      :hackney_pool.child_spec(:token_instance_fetcher, max_connections: pool_size),
       {Memory.Monitor, [memory_monitor_options, [name: memory_monitor_name]]},
       {CoinBalanceOnDemand.Supervisor, [json_rpc_named_arguments]},
       {TokenTotalSupplyOnDemand.Supervisor, []},